chore: Rename dynamo.ingress to dynamo.frontend
We now start HTTP+pre-process+router like this:
```
python -m dynamo.frontend [--http-port 8080]
```

This fits the naming we agreed on. Previously the dynamo.frontend namespace was taken up by the examples, but those are gone now.
