+++ components/frontend/src/dynamo/frontend/__main__.py
@@ -1,7 +1,7 @@
 #  SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 #  SPDX-License-Identifier: Apache-2.0
 
-from dynamo.ingress.main import main
+from dynamo.frontend.main import main
 
 if __name__ == "__main__":
     main()
