+++ components/frontend/src/dynamo/frontend/main.py
@@ -1,7 +1,7 @@
 #  SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 #  SPDX-License-Identifier: Apache-2.0
 
-# Usage: `python -m dynamo.ingress [args]`
+# Usage: `python -m dynamo.frontend [args]`
 #
 # Start a frontend node. This runs:
 # - OpenAI HTTP server.
@@ -24,18 +24,15 @@ def parse_args():
         formatter_class=argparse.RawTextHelpFormatter,  # To preserve multi-line help formatting
     )
     parser.add_argument(
+        "-i", "--interactive", action="store_true", help="Interactive text chat"
+    )
+    parser.add_argument(
         "--kv-cache-block-size", type=int, help="KV cache block size (u32)."
     )
     parser.add_argument(
         "--http-port", type=int, default=8080, help="HTTP port for the engine (u16)."
     )
-    flags = parser.parse_args()
-
-    kwargs = {"http_port": flags.http_port}
-    if flags.kv_cache_block_size is not None:
-        kwargs["kv_cache_block_size"] = flags.kv_cache_block_size
-
-    return kwargs
+    return parser.parse_args()
 
 
 async def async_main():
@@ -43,12 +40,18 @@ async def async_main():
     flags = parse_args()
 
     # out=dyn
-    e = EntrypointArgs(EngineType.Dynamic, **flags)
+    e = EntrypointArgs(
+        EngineType.Dynamic,
+        http_port=flags.http_port,
+        kv_cache_block_size=flags.kv_cache_block_size,
+    )
     engine = await make_engine(runtime, e)
 
-    # in=http
     try:
-        await run_input(runtime, "http", engine)
+        if flags.interactive:
+            await run_input(runtime, "text", engine)
+        else:
+            await run_input(runtime, "http", engine)
     except asyncio.exceptions.CancelledError:
         pass
